+++ php-src/Schemas/Php.php
@@ -26,7 +26,7 @@ class Php extends ASchema
         return null;
     }
 
-    public function getTimeout(): ?int
+    public function getTimeout(): ?float
     {
         return null;
     }
